+++ lib/containers/stivec.h
@@ -1,15 +1,15 @@
 #pragma once
 
-#include <stdint.h>
+#include <stddef.h>
 
 typedef struct {
-    uint32_t* ids;
-    uint32_t* elems;
-    uint32_t size;
+    size_t* ids;
+    size_t* elems;
+    size_t size;
 } stivec_t;
 
-stivec_t create_stivec(uint32_t max_size);
-uint32_t stivec_insert(stivec_t* stivec, uint32_t val);
-void stivec_remove(stivec_t* stivec, uint32_t id);
-uint32_t stivec_get(stivec_t* stivec, uint32_t id);
-uint32_t stivec_random_pop(stivec_t* stivec);
+stivec_t create_stivec(size_t max_size);
+size_t stivec_insert(stivec_t* stivec, size_t val);
+void stivec_remove(stivec_t* stivec, size_t id);
+size_t stivec_get(stivec_t* stivec, size_t id);
+size_t stivec_random_pop(stivec_t* stivec);
